Move TransactionInner and add get methods to Transaction
